+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 
 ## Unreleased
 
+### Changed
+
+- Divide `kamoshika.py` into several files
+- Remove `responce` (typo of "response") field from `sample.yml` because change `kamoshika.py` behavior not to refer the field
+
 ---
 
 ## [v0.6.0] - 2018/03/24